[FIX] React being loaded on the main bundle (#18597)
