+++ client/account/sidebarItems.js
@@ -1,7 +1,7 @@
 import { hasPermission } from '../../app/authorization/client';
 import { createTemplateForComponent } from '../reactAdapters';
 import { settings } from '../../app/settings';
-import { createSidebarItems } from '../components/basic/Sidebar';
+import { createSidebarItems } from '../helpers/createSidebarItems';
 
 createTemplateForComponent('accountFlex', () => import('./AccountSidebar'));
